plc4go: Cleanup + Added knxnetip_driver_test.go and knxnetip_parser_serializer_test.go for intial skipped test runs + Deduplicated Package names out of struct names + Converted consts to camel case + Added timeout for timeout error + removed package shadowing on modbus/Reader.go + fixed several linter issues

+++ plc4go/internal/plc4go/spi/utils/CastUtils.go
@@ -70,7 +70,7 @@ func ByteArrayToInt8Array(input []byte) []int8 {
 func ByteArrayToUint8Array(input []byte) []uint8 {
 	output := make([]uint8, len(input))
 	for i, _val := range input {
-		output[i] = uint8(_val)
+		output[i] = _val
 	}
 	return output
 }
